// apps/serverless-admin/src/lib/schemas.ts
export const layerRefSchema = z.object({
  name: z.string(),
  version: z.number(),
  arn: z.string().optional(),
})

export const functionVersionSchema = z.object({
  version: z.string(),
  versionNumber: z.number().optional(),
  codeSha256: z.string().optional(),
  codeArtifact: codeArtifactSchema.nullish(),
  createdAt: z.string().optional(),
})

export const functionSchema = z.object({
  id: z.string(),
  accountId: z.string(),
  namespace: z.string(),
  region: z.string(),
  name: z.string(),
  functionArn: z.string().optional(),
  packageType: z.string(),
  imageUri: z.string().optional(),
  runtime: z.string().optional(),
  handler: z.string().optional(),
  architecture: z.string().optional(),
  runtimeMode: z.string().optional(),
  memorySize: z.number().optional(),
  timeout: z.number().optional(),
  reservedConcurrency: z.number().optional(),
  layers: z.array(layerRefSchema).optional(),
  env: z.record(z.string(), z.string()).optional(),
  state: z.string(),
  createdAt: z.string(),
  updatedAt: z.string(),
  version: functionVersionSchema.optional(),
})
// ...
export type FunctionEntity = z.infer<typeof functionSchema>